+++ culture-explorer-kids/src/__pycache__/culture_logic.cpython-39.pyc
